requirements.txt
logging
openai-agents
logfire
datetime
azure-search-documents>=11.4.0
azure-core>=1.29.0
openai>=1.3.5
streamlit>=1.32.0
azure-search-documents>=11.4.0
azure-core>=1.29.0
openai>=1.3.5
google-search-results
nest_asyncio>=1.5.8
sqlparse
pyodbc
azure-cosmos
azure-identity 
azure-keyvault-secrets